+++ CHANGELOG.md
@@ -5,12 +5,15 @@
 - Add screenshot support for images via the URL (if the image's URL has an "screenshot" in it, it will be displayed as a screenshot)
 
 - Add footer links
+  - And icorporate the "Source Code" and "Made by" links into the footer links
 
 - Add styles for check boxes
 
 - Add `markdown-it` plugins support for better markdown parsing
 
 ### Cleanup(s)
+
+- Slighly darker background color for dark mode
 
 - Make a more destinctive style for descriptions in the article's header