+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add screenshot support for images via the URL (if the image's URL has an "screenshot" in it, it will be displayed as a screenshot)
 
 - Add footer links
+
   - And icorporate the "Source Code" and "Made by" links into the footer links
 
 - Add styles for check boxes